backend/requirements-worker.txt
# Dependencias pesadas para calculos (GitHub Actions)
pandas==2.3.3
numpy==2.4.0
scikit-learn==1.8.0
scipy==1.16.3
joblib==1.3.2

# Base de datos y utilidades
sqlalchemy==2.0.45
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# HTTP clients for data fetching
httpx==0.26.0
aiohttp==3.13.2

# Core FastAPI (for shared DTOs and entities)
pydantic==2.12.5
pydantic-settings==2.12.0

# Timezone handling
pytz==2024.1

# Task scheduling (for worker script)
apscheduler==3.11.2